+++ Changelog.md
@@ -14,6 +14,11 @@ Der entsprechende Abschnitt des Changelogs wird auch jeweils in die [Releasenote
 
 ## Unreleased
 
+
+### Neu
+
+- "gzip" encoding akzeptieren
+
 ## [12.16.1] - 17.12.2025
 
 ### Änderung